fix: Start to split up the native code into different files for better navigation and maintenance.
