+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/json_helper.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "json.h"
+
+namespace playeveryware::eos::json_helpers
+{
+    
+
+}
