+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/json_helpers.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace playeveryware::eos::json_helpers
+{
+    
+}
