+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/string_helpers.cpp
@@ -1,0 +1,7 @@
+#include <pch.h>
+#include <string_helpers.h>
+
+namespace playeveryware::eos::string_helpers
+{
+    
+}
